move tests to more appropriate location
